pyproject.toml
[tool.poetry]
name = "project15"
version = "0.1.0"
description = ""
authors = ["Dmitry Stolyarov <leshiy74@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "6.*"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
